Stop FBP service of the session when the window is closed
When implementing the login system, it will be able to let the services
running when the window of the browser is closed.

Without this patch the guest session may be renewed and the coder will
not be able to stop the service. Then, the service will stay haging out
until the next restart of the system.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>