Fix: tmp files exists when resyncing a repository
If a session is closed in the middle of a sync process the progress will be
interruped. When this closed session restores the sync may fail since the tmp
clone folder will be occupied.

With this patch the api will try to remove the tmp folder first, before
continuing the clone process.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>