Verify what nodejs name is
Linux distros may change the name of nodejs binary file,it was found two names: "
node" or "nodejs".
The script has to identify which name of the nodejs binary is being used before
continuing.

Also, depending on the distro sh files may change and it is more secure to use
/bin/bash than /bin/sh. For compatibility purposes

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>