Detect when file was changed and asks to be saved
Now whenever the file changes the save icon will be shown as enabled and
it will put a * char in front of the filename.

If the file was changed and the page is closed, it will show a dialog asking to
warning that the file was changed and it may be saved.

This patch also fixes some syntaxes issues.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>