fix: remove unneeded error check
It was caught some warn messages that were misleading the Devapp to identify
unneeded error.

The warn message comes from Soletta and it was reported to be
checked.

In any case if an wrong fbp code is sent to the server it will cause error
when trying to generate the svg file from it.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>