Replace method to parse the json of journalctl
Since the journalctl does not return an array of each of its lines, let's
parse it according to the json block provided.

Now when there is a parsing error of a single block, it will print error in
console showing what line was not able to parse and it will proceed to get
another new inputs that is valid. Avoid getting blank screen if a single
line is with problem.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>