Fix: nodejs pushing unreadable messages to journal
Node is pushing unreadable messages on journalctl containing only
array of numbers such as:
"[ 27, 91, 48, 109, 112...]"

Since after some tests this is the only message the node send to
the journal, it was decided to remove any message coming from the
syslog called node.

This will no affect any warn/error messages that comes from the
server.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>