syntax highlight: fix non-blue uppercase var name
When the nodetype consists in upper case letter it wasn't coloring
the whole name.

This patch also fix float number that was having different colors after the
point. Such as: 1.2

The 2 was be pink and the 1 was blue.

Also, this patch contains changes in var name to be more human readable.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>