Run FBP program in repo instead of copying stucture to tmp
The running method was focused in building a run environment before running
the project.

When DECLARE was created, Soletta Dev-App started creating a
building environment wrongly when DECLARE was used with relative PATH.

With this patch Soletta Dev-App will run the project directly in the
right PATH of the repo, making DECLARE working with Relative paths.

Fixes the issue: Devapp wasn't finding the external FBP when using DECLARE in a
relative path.