Fix: Do not use display:none attribute for cheat-sheet div
In Firefox only, cheat-sheet was not loading properly.

After a research it was discovered a well-known bug that if combined
display:none in a div that contains an iframe then Firefox will fail to load
properly the iframe.

This is a workaround solution for a bug on Firefox only:
https://github.com/desandro/get-size/#firefox-hidden-iframe-bug

Reported bug:
https://bugzilla.mozilla.org/show_bug.cgi?id=548397

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>