editor: When creating new file check if it already exists
Check if the name of the new file exists in the same folder of the repository
if it exists it will pop a dialog asking if it can be override.

If the user click yes it will erase the existed file.
If the user click no it will cancel the creation.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>