client: Add a way to reopen the run dialog
Add a menu entry to allow users to reopen the run dialog, where they can
check the std output of the fbp currently being executed.

The run dialog has a close button, however there wasn't a way to
reopen it - Users had to stop, then rerun the application in order
to bring this dialog again.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@intel.com>